requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
google-cloud-texttospeech>=2.16.0
